Update multiple-projects.md with more details how to set up multiple projects
More details how to set up multiple projects.

This is actually achieved with Træfik so the title of the doc page is renamed to "Running Multiple Projects with Træfik"
An image is included for users to quickly get an idea how it works.
An example is used to describe all steps required to get a multi-project environment running. After that point users should be able to achieve a level of understanding to make their own adjustments.